Extract insertBr() functionality into a proper "InsertLineBreak" editor command, so that behavior can be standardized across TinyMCE and e.g. invoked from an external toolbar button or menu item dedicated to achieving the equivalent behavior to typing shift-enter.
